+++ Source/Camera.h
@@ -48,7 +48,7 @@ namespace Theodore {
 		void SetRenderTexture(FrameBuffer* texture);
 
 	private:
-		virtual void Update(double deltaTime) override;
+		virtual void Update(float deltaTime) override;
 		virtual void Render() override;
 		virtual bool CompareEquality(const Object& rhs) const override;
 		virtual bool Destroy() override;
