+++ Source/Gizmo.h
@@ -20,7 +20,7 @@ namespace Theodore {
 		virtual ~Gizmo();
 
 	private:
-		virtual void Update(double deltaTime) override;
+		virtual void Update(float deltaTime) override;
 		virtual void Render() override;
 		virtual bool CompareEquality(const Object& rhs) const override;
 		virtual bool Destroy() override;
