+++ Source/MeshRenderer.h
@@ -23,7 +23,7 @@ namespace Theodore {
 		Mesh* GetMesh() const;
 
 	private:
-		virtual void Update(double deltaTime) override;
+		virtual void Update(float deltaTime) override;
 		virtual void Render() override;
 		virtual bool CompareEquality(const Object& rhs) const override;
 		virtual bool Destroy() override;
