+++ Source/Transform.h
@@ -84,7 +84,7 @@ namespace Theodore {
 		Matrix4x4 GetLocalMatrix() const;
 		Matrix4x4 GetParentLocalMatrix() const;
 
-		virtual void Update(double deltaTime) override;
+		virtual void Update(float deltaTime) override;
 		virtual void Render() override;
 		virtual bool CompareEquality(const Object& rhs) const override;
 		virtual bool Destroy() override;
